+++ main.py
@@ -8,7 +8,6 @@ If the code doesn't run when cloning from GitHub, the data folder is assumed to 
 """
 
 from methods.process import Experiment
-from methods.evaluate import Evaluate
 
 
 def main():
@@ -19,20 +18,24 @@ def main():
     """
     experiment = Experiment(
         time_experiments=True,
-        verbose=True)
+        verbose=True,
+        debug=False  # This cuts the data by a debug factor.
+    )
+    experiment.debug_cutoff = 0.1
     # Comment models in order to exclude from the experiment...
     experiment.models = [
         'naive-bayes',
         'svm',
         'logistic',
-        # 'random-forest',  # This lovely method takes hours -> goes in the backlog
-        # 'word2vec',
         # Neural
         'cnn',
-        'lstm'
+        'lstm',
+        'gru'
     ]
-    # experiment.perform_many_experiments(save_pipes=True)
-    experiment.perform_single_experiment(pipeline_model='word2vec', save_pipe=True)
+    # experiment.perform_many_experiments(save_pipes=False, load_pipes=False)
+    # experiment.perform_single_experiment(pipeline_model='svm', save_pipe=False, load_pipe=False)
+    experiment.cross_validate_experiments()
+    # print(experiment.labels.shape)
 
 
 if __name__ == '__main__':